requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
Django==4.2.17
django-allauth==65.3.0
djangorestframework==3.15.2
flake8==7.1.1
idna==3.10
kombu==5.4.2
mccabe==0.7.0
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
redis==5.2.1
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
